Build and run
This gets the package to fully build and will successfully run
Trenchbroom.

Defaulting to only allowing access to the user's document folder. This
seems like a sane safe default and user's can easily add access to other
folders using:
`flatpak override --filesystem=/path/to/folder
com.kristianduske.TrenchBroom`

Also added `--persist` for the .TrenchBroom folder otherwise the app
will not behave correctly.

This adds pandoc because it is needed by trenchbroom's build process but
it uses the official binary instead of building it from source to speed
up the build process. It is removed during the cleanup phase.